requirements.txt
aiohttp
python-dotenv
pydantic
tenacity
